Harden scraper package checks
