+++ scraper/Dockerfile
@@ -2,6 +2,6 @@ FROM node:20-alpine
 
 WORKDIR /app
 COPY package*.json ./
-RUN npm install
+RUN npm ci
 COPY . .
 CMD ["npx", "tsx", "src/scrape.ts"]
